back-end/app/pyproject.toml
[project]
name = "Xplore-Backend"
version = "0.1.0"
description = "Backend API for Xplore."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "langchain>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-google-genai>=2.1.5",
    "passlib[bcrypt]>=1.7.4",
    "pip-tools>=7.4.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "redis[hiredis]>=6.2.0",
    "sqlmodel>=0.0.24",
]

[dependency-groups]
dev = [
    "ruff>=0.12.4",
]
